requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
sqlmodel
psycopg2-binary
sqlalchemy
httpx
PyJWT